Rename eye_model/blocking.py to eye_model/base.py
